Refactor: NewModal 네이밍 Modal로 수정
