+++ src/components/LoginModal.tsx
@@ -1,7 +1,7 @@
 import NaverLoginButton from "@/assets/svg/naver-login-btn.svg";
 import KakaoLoginButton from "@/assets/svg/kakao-login-btn.svg";
 import GoogleLoginButton from "@/assets/svg/google-login-btn.svg";
-import NewModal from "@/components/common/NewModal";
+import Modal from "@/components/common/Modal";
 
 interface Props {
   isOpen: boolean;
@@ -10,9 +10,9 @@ interface Props {
 
 const LoginModal = ({ isOpen, onClose }: Props) => {
   return (
-    <NewModal isOpen={isOpen} onClose={onClose}>
-      <NewModal.Header>로그인</NewModal.Header>
-      <NewModal.Body>
+    <Modal isOpen={isOpen} onClose={onClose}>
+      <Modal.Header>로그인</Modal.Header>
+      <Modal.Body>
         <div className="flex flex-col items-center gap-3">
           <a href="/oauth2/authorization/google">
             <GoogleLoginButton />
@@ -24,8 +24,8 @@ const LoginModal = ({ isOpen, onClose }: Props) => {
             <NaverLoginButton />
           </a>
         </div>
-      </NewModal.Body>
-    </NewModal>
+      </Modal.Body>
+    </Modal>
   );
 };
